+++ infra/docker/order/initdb.d/V01__create_orders.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS orders (
     symbol VARCHAR(32) NOT NULL,
     side VARCHAR(10) NOT NULL,
     type VARCHAR(10) NOT NULL,
-    qty BIGINT NOT NULL,
+    qty NUMERIC(19,4) NOT NULL,
     limit_price NUMERIC(18,8),
     status VARCHAR(32) NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT now() NOT NULL,
